refactor: move validate_quality to cli module
